+++ decoder.go
@@ -14,13 +14,13 @@ import "C"
 
 // Decoder is used for encoding data
 type Decoder struct {
-	mDecoder C.kodo_rlnc_decoder_t
+	mDecoder *C.krlnc_decoder_t
 }
 
-// freeDecoder deallocates and release the memory consumed by an decoder
+// deleteDecoder deallocates and release the memory consumed by an decoder
 // @param decoder The decoder which should be deallocated
-func freeDecoder(decoder *Decoder) {
-	C.kodo_rlnc_decoder_destruct(decoder.mDecoder)
+func deleteDecoder(decoder *Decoder) {
+	C.krlnc_delete_decoder(decoder.mDecoder)
 }
 
 // PayloadSize returns the payload size of an decoder, which is the
@@ -28,7 +28,7 @@ func freeDecoder(decoder *Decoder) {
 // @param decoder The decoder to query.
 // @return The payload size in bytes
 func (decoder *Decoder) PayloadSize() uint32 {
-	return uint32(C.kodo_rlnc_decoder_payload_size(decoder.mDecoder))
+	return uint32(C.krlnc_decoder_payload_size(decoder.mDecoder))
 }
 
 // ReadPayload reads symbol from the given payload buffer.
@@ -38,7 +38,7 @@ func (decoder *Decoder) PayloadSize() uint32 {
 //        so it cannot be reused. If the payload is needed at several places,
 //        make sure to keep a copy of the original payload.
 func (decoder *Decoder) ReadPayload(payload *[]uint8) {
-	C.kodo_rlnc_decoder_read_payload(
+	C.krlnc_decoder_read_payload(
 		decoder.mDecoder, (*C.uint8_t)(&(*payload)[0]))
 }
 
@@ -47,21 +47,21 @@ func (decoder *Decoder) ReadPayload(payload *[]uint8) {
 // @return The block size, i.e. the total size in bytes that this decoder
 //         operates on.
 func (decoder *Decoder) BlockSize() uint32 {
-	return uint32(C.kodo_rlnc_decoder_block_size(decoder.mDecoder))
+	return uint32(C.krlnc_decoder_block_size(decoder.mDecoder))
 }
 
 // SymbolSize returns the symbol size of the decoder.
 // @param decoder The decoder to check
 // @return The size of a symbol in bytes
 func (decoder *Decoder) SymbolSize() uint32 {
-	return uint32(C.kodo_rlnc_decoder_symbol_size(decoder.mDecoder))
+	return uint32(C.krlnc_decoder_symbol_size(decoder.mDecoder))
 }
 
 // Symbols returns the number of symbols in a block.
 // @param decoder The decoder to check
 // @return The number of symbols
 func (decoder *Decoder) Symbols() uint32 {
-	return uint32(C.kodo_rlnc_decoder_symbols(decoder.mDecoder))
+	return uint32(C.krlnc_decoder_symbols(decoder.mDecoder))
 }
 
 // SetMutableSymbols specifies the data buffer where the decoder
@@ -71,7 +71,7 @@ func (decoder *Decoder) Symbols() uint32 {
 // @param data The buffer that should contain the decoded symbols
 // @param size The size of the buffer to be decoded
 func (decoder *Decoder) SetMutableSymbols(data *[]uint8) {
-	C.kodo_rlnc_decoder_set_mutable_symbols(
+	C.krlnc_decoder_set_mutable_symbols(
 		decoder.mDecoder, (*C.uint8_t)(&(*data)[0]), C.uint32_t(len(*data)))
 }
 
@@ -79,7 +79,7 @@ func (decoder *Decoder) SetMutableSymbols(data *[]uint8) {
 // @param decoder The decoder to query
 // @return true if the decoding is complete, otherwise false
 func (decoder *Decoder) IsComplete() bool {
-	return C.kodo_rlnc_decoder_is_complete(decoder.mDecoder) != 0
+	return C.krlnc_decoder_is_complete(decoder.mDecoder) != 0
 }
 
 // Rank returns the rank of a decoder indicates how many symbols have
@@ -87,5 +87,5 @@ func (decoder *Decoder) IsComplete() bool {
 // @param decoder The decoder to query
 // @return The rank of the decoder
 func (decoder *Decoder) Rank() uint32 {
-	return uint32(C.kodo_rlnc_decoder_rank(decoder.mDecoder))
+	return uint32(C.krlnc_decoder_rank(decoder.mDecoder))
 }