+++ encoder.go
@@ -14,13 +14,13 @@ import "C"
 
 // Encoder is used for encoding data
 type Encoder struct {
-	mEncoder C.kodo_rlnc_encoder_t
+	mEncoder *C.krlnc_encoder_t
 }
 
-// freeEncoder deallocates and release the memory consumed by an encoder
+// deleteEncoder deallocates and release the memory consumed by an encoder
 // @param encoder The encoder which should be deallocated
-func freeEncoder(encoder *Encoder) {
-	C.kodo_rlnc_encoder_destruct(encoder.mEncoder)
+func deleteEncoder(encoder *Encoder) {
+	C.krlnc_delete_encoder(encoder.mEncoder)
 }
 
 // PayloadSize returns the payload size of an encoder, which is the
@@ -28,7 +28,7 @@ func freeEncoder(encoder *Encoder) {
 // @param encoder The encoder to query.
 // @return The payload size in bytes
 func (encoder *Encoder) PayloadSize() uint32 {
-	return uint32(C.kodo_rlnc_encoder_payload_size(encoder.mEncoder))
+	return uint32(C.krlnc_encoder_payload_size(encoder.mEncoder))
 }
 
 // WritePayload writes a symbol into the provided payload buffer.
@@ -36,7 +36,7 @@ func (encoder *Encoder) PayloadSize() uint32 {
 // @param payload The buffer which should contain the symbol.
 // @return The total bytes used from the payload buffer
 func (encoder *Encoder) WritePayload(payload *[]uint8) uint32 {
-	return uint32(C.kodo_rlnc_encoder_write_payload(
+	return uint32(C.krlnc_encoder_write_payload(
 		encoder.mEncoder, (*C.uint8_t)(&(*payload)[0])))
 }
 
@@ -45,21 +45,21 @@ func (encoder *Encoder) WritePayload(payload *[]uint8) uint32 {
 // @return The block size, i.e. the total size in bytes that this encoder
 //         operates on.
 func (encoder *Encoder) BlockSize() uint32 {
-	return uint32(C.kodo_rlnc_encoder_block_size(encoder.mEncoder))
+	return uint32(C.krlnc_encoder_block_size(encoder.mEncoder))
 }
 
 // SymbolSize returns the symbol size of the encoder.
 // @param encoder The encoder to check
 // @return The size of a symbol in bytes
 func (encoder *Encoder) SymbolSize() uint32 {
-	return uint32(C.kodo_rlnc_encoder_symbol_size(encoder.mEncoder))
+	return uint32(C.krlnc_encoder_symbol_size(encoder.mEncoder))
 }
 
 // Symbols returns the number of symbols in a block.
 // @param encoder The encoder to check
 // @return The number of symbols
 func (encoder *Encoder) Symbols() uint32 {
-	return uint32(C.kodo_rlnc_encoder_symbols(encoder.mEncoder))
+	return uint32(C.krlnc_encoder_symbols(encoder.mEncoder))
 }
 
 // SetConstSymbols specifies the source data for all symbols.
@@ -68,7 +68,7 @@ func (encoder *Encoder) Symbols() uint32 {
 // @param data The buffer containing the data to be encoded
 // @param size The size of the buffer to be encoded
 func (encoder *Encoder) SetConstSymbols(data *[]uint8) {
-	C.kodo_rlnc_encoder_set_const_symbols(
+	C.krlnc_encoder_set_const_symbols(
 		encoder.mEncoder, (*C.uint8_t)(&(*data)[0]), C.uint32_t(len(*data)))
 }
 
@@ -77,17 +77,17 @@ func (encoder *Encoder) SetConstSymbols(data *[]uint8) {
 // @param encoder The encoder
 // @return Non-zero if the encoder is in the systematic mode, otherwise 0
 func (encoder *Encoder) IsSystematicOn() bool {
-	return C.kodo_rlnc_is_systematic_on(encoder.mEncoder) != 0
+	return C.krlnc_is_systematic_on(encoder.mEncoder) != 0
 }
 
 // SetSystematicOn switches the systematic encoding on
 // @param encoder The encoder
 func (encoder *Encoder) SetSystematicOn() {
-	C.kodo_rlnc_set_systematic_on(encoder.mEncoder)
+	C.krlnc_set_systematic_on(encoder.mEncoder)
 }
 
 // SetSystematicOff switches the systematic encoding off
 // @param encoder The encoder
 func (encoder *Encoder) SetSystematicOff() {
-	C.kodo_rlnc_encoder_set_systematic_off(encoder.mEncoder)
+	C.krlnc_encoder_set_systematic_off(encoder.mEncoder)
 }