+++ NEWS.rst
@@ -6,7 +6,9 @@ every change, see the Git log.
 
 Latest
 ------
-* tbd
+* Major: Renamed ``free*`` methods to ``delete*`` methods.
+* Minor: Enabled binary16 field.
+* Major: Use kodo-rlnc-c version 2.
 
 1.0.0
 -----